Move testbed/correctness from collector core to collector contrib
