+++ testbed/Makefile
@@ -19,3 +19,19 @@ list-stability-tests:
 .PHONY: run-stability-tests
 run-stability-tests:
 	TESTCASE_DURATION=1h TEST_ARGS="$${TEST_ARGS} -timeout 70m" TESTS_DIR=stabilitytests ./runtests.sh
+
+.PHONY: list-correctness-metrics-tests
+list-correctness-metrics-tests:
+	RUN_TESTBED=1 $(GOTEST) -v ./correctnesstests/metrics --test.list '.*' | grep "^Test"
+
+.PHONY: run-correctness-metrics-tests
+run-correctness-metrics-tests:
+	TESTS_DIR=correctnesstests/metrics ./runtests.sh
+
+.PHONY: list-correctness-traces-tests
+list-correctness-traces-tests:
+	RUN_TESTBED=1 $(GOTEST) -v ./correctnesstests/traces --test.list '.*' | grep "^Test"
+
+.PHONY: run-correctness-traces-tests
+run-correctness-traces-tests:
+	TESTS_DIR=correctnesstests/metrics ./runtests.sh
