+++ testbed/correctnesstests/metrics/testdata/pict_input_metrics_pipeline.txt
