+++ testbed/correctnesstests/traces/testdata/generated_pict_pairs_traces_pipeline.txt
