+++ testbed/correctnesstests/traces/testdata/pict_input_traces_pipeline.txt
